requirements.txt
feedparser==6.0.10
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
scikit-learn==1.2.2
requests==2.31.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
diffusers==0.19.0
transformers==4.31.0
accelerate==0.20.3
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.100.0
boto3==1.28.0
markdown==3.4.3
pillow==10.0.1
